feat(extras): Volume.mount(read_only=True)
Add a ``read_only`` knob to ``Volume.mount``. When enabled, the FUSE
mount is brought up with juicefs ``--read-only`` so writes return EROFS
and the writeback queue isn't engaged. The metadata engine still starts
(Redis loads the index) but no chunks ever upload.

Use case: expose a prior Volume's state read-only — multiple readers
can attach to the same index, no concurrent-writer hazards.
